Fruit of Might

+++ src/main/java/kamkeel/npcdbc/api/effect/IBonusHandler.java
@@ -4,13 +4,13 @@ import noppes.npcs.api.entity.IPlayer;
 
 public interface IBonusHandler {
 
-    IPlayerBonus createBonus(String name, byte str, byte dex, byte wil);
+    IPlayerBonus createBonus(String name, float str, float dex, float wil);
 
     boolean hasBonus(IPlayer player, String name);
 
     boolean hasBonus(IPlayer player, IPlayerBonus bonus);
 
-    void applyBonus(IPlayer player, String name, byte str, byte dex, byte wil);
+    void applyBonus(IPlayer player, String name, float str, float dex, float wil);
 
     void applyBonus(IPlayer player, IPlayerBonus bonus);
